handle order in Checkout component
